Inital commit
Right now, ma.py supports reading in an optional country name, in case someone wants to test his knowldge about a certain country.
Not case-sensitive, and uses pycountry to get the list of updated countries.
Not uses argparse's 'choices', as this leads to a huge list which is always displayed on screen with all country names; instead, uses the 'type' method, which enables checking if the country name given by the user is valid (according to pycountry) or not.