Added support for countries with long names
Arguments of countries with long names did not parse well. The function check_valid_country was modified to support such countries, like 'south africa', 'costa rica' etc.
Moreover, edge-cases are also supported; if a partial country name is added (for example, just 'south'), then the program will respond with a different argparse exception, to indicate that the user probably forgot to add the rest of the name. This part is implemented as part of the Mapy class.